requirements.txt
# ===== Core runtime =====
numpy==2.0.0               # in linea con l’ambiente attuale
pandas==2.2.3
openpyxl>=3.1.2            # lettura Excel per curve batteria
pyyaml>=6.0


requests>=2.31.0
pytz>=2023.3

# ===== Microgrid simulator =====
python-microgrid-custom @ git+https://github.com/ferro-gianluca-29/python-microgrid.git@65e29a0866e3511b35b17bdf5b3650d83f083819
gymnasium>=0.29.1          # drop-in replacement di gym, compatibile con pymgrid
cvxpy>=1.3.0
ecos>=2.0.12               # solver ECOS/ECOS_BB usato come fallback per l'MPC
gurobipy>=10.0.0           # binding Python per GUROBI (richiede installazione/ licenza native)

# ===== Reinforcement learning (RL_AGENT) =====
stable-baselines3>=2.0.0   # PPO, DummyVecEnv, ecc.
torch>=2.0.0               # backend per stable-baselines3 (PyTorch)
tensorboard>=2.16.0       # logging per TensorBoard (RL_AGENT)


# ===== GUI opzionale =====
setuptools==80.9.0        # pandasgui richiede pkg_resources (rimosso in setuptools 82+)
pandasgui==0.2.15
PyQt5==5.15.11
PyQtWebEngine==5.15.7

imitation>=0.3.1
matplotlib>=3.6
tqdm>=4.64
